Stuart Buchanan:
Add a  Mainair Flash 2 microlight to the hangar.

"The weather in Scotland the last couple of months has been somewhat
challenging for learning to fly, particularly in a microlight which cannot
handle high winds as well as a GA aircraft. The headwind limit on my
machine is about 25mph, and the crosswind limit is a 6mph!"